src/module_search/todo_requirements.txt
Pillow
torchvision
openai-clip 
torch